Separate out service day logic from DateTime and put in mjolnir since it is only used when bulding data. This removes some boost functionality from the regular build. Separate servicedays tests as well.
