+++ src/baldr/datetime.cc
@@ -58,18 +58,6 @@ const tz_db_t& get_tz_db() {
   return tz_db;
 }
 
-// default testing date and time is the next Tues @ 08:00
-std::string get_testing_date_time() {
-  auto tz = get_tz_db().from_index(get_tz_db().to_index("America/New_York"));
-  boost::gregorian::date d = get_formatted_date(iso_date_time(tz));
-
-  while (d.day_of_week() != boost::date_time::Tuesday) {
-    d += boost::gregorian::days(1);
-  }
-
-  return to_iso_extended_string(d) + "T08:00";
-}
-
 // get a formatted date.
 boost::gregorian::date get_formatted_date(const std::string& date) {
   boost::gregorian::date d;
@@ -116,146 +104,6 @@ boost::local_time::local_date_time get_ldt(const boost::gregorian::date& date,
     }
   }
   return in_local_time;
-}
-
-// Get service days
-// Start from the tile_date or start date to end date or 60 days (whichever is less)
-// start_date will be updated to the tile creation date if the start date is in the past
-// set the bits based on the dow.
-uint64_t get_service_days(boost::gregorian::date& start_date,
-                          boost::gregorian::date& end_date,
-                          const uint32_t tile_date,
-                          const uint32_t dow_mask) {
-
-  boost::gregorian::date tile_header_date = pivot_date_ + boost::gregorian::days(tile_date);
-
-  // if our start date is more than 60 days out, reject.
-  if (start_date > (tile_header_date + boost::gregorian::days(59))) {
-    return 0;
-  }
-
-  if (start_date <= tile_header_date && tile_header_date <= end_date) {
-    start_date = tile_header_date;
-  } else if (tile_header_date > end_date) { // reject.
-    return 0;
-  }
-
-  // only support 60 days out.  (59 days and include the end_date = 60)
-  boost::gregorian::date enddate = tile_header_date + boost::gregorian::days(59);
-
-  if (enddate <= end_date) {
-    end_date = enddate;
-  }
-
-  uint32_t days = 0;
-  boost::gregorian::day_iterator itr(tile_header_date + boost::gregorian::days(days));
-
-  uint32_t x = days;
-  uint64_t bit_set = 0;
-
-  // TODO: we dont have to loop over all days, we could take the week mask, shift it by the start
-  // date then use that weekly mask and shift it how ever many times a week fits into the 60 day
-  // interval loop over all the days
-  while (itr <= end_date) {
-
-    // figure out what day of the week we are at
-    uint8_t dow;
-    switch ((*itr).day_of_week().as_enum()) {
-      case boost::date_time::Sunday:
-        dow = kSunday;
-        break;
-      case boost::date_time::Monday:
-        dow = kMonday;
-        break;
-      case boost::date_time::Tuesday:
-        dow = kTuesday;
-        break;
-      case boost::date_time::Wednesday:
-        dow = kWednesday;
-        break;
-      case boost::date_time::Thursday:
-        dow = kThursday;
-        break;
-      case boost::date_time::Friday:
-        dow = kFriday;
-        break;
-      case boost::date_time::Saturday:
-        dow = kSaturday;
-        break;
-    }
-
-    // were we supposed to be on for this day?
-    // and are we at or after the start date?
-    if ((dow_mask & dow) && (itr >= start_date)) {
-      bit_set |= static_cast<uint64_t>(1) << x;
-    }
-
-    ++itr;
-    ++x;
-  }
-  return bit_set;
-}
-
-// add a service day to the days if it is in range.
-uint64_t add_service_day(const uint64_t& days,
-                         const boost::gregorian::date& end_date,
-                         const uint32_t tile_date,
-                         const boost::gregorian::date& added_date) {
-
-  // adding a service day must start at the tile header date.
-  boost::gregorian::date start_date = pivot_date_ + boost::gregorian::days(tile_date);
-  boost::gregorian::date enddate = start_date + boost::gregorian::days(59);
-
-  if (enddate > end_date) {
-    enddate = end_date;
-  }
-
-  if (start_date <= added_date && added_date <= enddate) {
-    boost::gregorian::date_period range(start_date, added_date);
-    uint32_t length = range.length().days();
-    return days | (static_cast<uint64_t>(1) << length);
-  }
-  return days;
-}
-
-// remove a service day to the days if it is in range.
-uint64_t remove_service_day(const uint64_t& days,
-                            const boost::gregorian::date& end_date,
-                            const uint32_t tile_date,
-                            const boost::gregorian::date& removed_date) {
-
-  // removing a service day must start at the tile header date.
-  boost::gregorian::date start_date = pivot_date_ + boost::gregorian::days(tile_date);
-  boost::gregorian::date enddate = start_date + boost::gregorian::days(59);
-
-  if (enddate > end_date) {
-    enddate = end_date;
-  }
-
-  if (start_date <= removed_date && removed_date <= enddate) {
-    boost::gregorian::date_period range(start_date, removed_date);
-    uint32_t length = range.length().days();
-    return ~((~days) | (static_cast<uint64_t>(1) << length));
-  }
-  return days;
-}
-
-// check if service is available for a date.
-bool is_service_available(const uint64_t days,
-                          const uint32_t start_date,
-                          const uint32_t date,
-                          const uint32_t end_date) {
-
-  if (start_date <= date && date <= end_date) {
-    boost::gregorian::date start = pivot_date_ + boost::gregorian::days(start_date);
-    boost::gregorian::date d = pivot_date_ + boost::gregorian::days(date);
-
-    std::bitset<64> bit_set(days);
-    boost::gregorian::date_period range(start, d);
-    uint32_t length = range.length().days();
-    return bit_set.test(length);
-  }
-  return false;
 }
 
 // Get the number of days that have elapsed from the pivot date for the inputed date.
